Removed watchdog capability. Now simply injects hosts.
