+++ dockerfiles/logstash/.dockerignore
@@ -1,0 +1,3 @@
+*.pyc
+*.md
+Dockerfile
