+++ dockerfiles/logstash/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+python -u /opt/logstash/inject_config.py
+exec logstash $LOGSTASH_ARGS -f /opt/logstash/logstash.conf --allow-env
