--- dockerfiles/logstash/supervisord.conf
@@ -1,18 +1,0 @@
-[supervisord]
-nodaemon=true
-
-[program:elastic_watchdog]
-command=sh -c 'python -u /opt/logstash/elastic_watchdog.py; kill 1'
-redirect_stderr=true
-stdout_logfile=/dev/stdout
-stdout_logfile_maxbytes=0
-auto_start=true
-
-[program:logstash]
-#for logstsash 5
-#command=sh -c 'sleep 5 && logstash $LOGSTASH_ARGS -f /opt/logstash/logstash.conf -r; kill 1'
-command=sh -c 'sleep 5 && logstash $LOGSTASH_ARGS -f /opt/logstash/logstash.conf --allow-env --auto-reload; kill 1'
-redirect_stderr=true
-stdout_logfile=/dev/stdout
-stdout_logfile_maxbytes=0
-auto_start=true
